requirements.txt
tensorflow>=2.8
numpy>=1.26.4
pillow>=11.3.0
scikit-learn>=1.3.2
tensorflow-macos>=2.9.0
tensorflow-metal>=0.5.0
matplotlib>=3.10.3
seaborn>=0.13.2
tqdm>=4.67.1
tf2onnx>=1.16.1
coremltools>=8.3.0

